+++ Source/LostConnection/Statuses/Ailments/IrradiationAilment.h
@@ -34,7 +34,7 @@ private:
 private:
 	FString getStatusName() const override;
 
-	SIZE_T getActiveStatusesCount() const override;
+	int32 getActiveStatusesCount() const override;
 
 private:
 	void GetLifetimeReplicatedProps(TArray<FLifetimeProperty>& OutLifetimeProps) const override;
